Add files via upload
This is the zipped project folder of assignment #3 which already includes all of the necessary directories and their respective/corresponding files. Simply extract the downloaded the zipped folder and open the folder on your IDE. As long as you have XAMPP control panel and the database test with the proper tables ready via the phpmyadmin web client, you're pretty much ready. All you need to do is the command npm start in the Backend directory and npm run dev in the Frontend directory via the terminal. All in all, this is a big shortcut to get the web app running.